requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.0.0
pillow>=10.0.0
psutil>=5.9.0

# Memos/Pensieve (install separately if not available)
# memos>=0.1.0

# VLM optimization
imagehash>=4.3.1
requests>=2.31.0
numpy>=1.24.0
aiohttp>=3.8.0

# Pensieve integration dependencies
websockets>=15.0.0
psycopg2-binary>=2.9.0
sentence-transformers>=2.0.0

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# For testing
playwright>=1.40.0